requirements.txt
fastapi
uvicorn
line-bot-sdk
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
Pillow
httpx
aiofiles
google-cloud-vision
python-multipart
